requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
channels==4.0.0
channels-redis==4.2.0
Django==5.0.3
django-jazzmin==2.6.1
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
msgpack==1.0.8
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
